chore: rename vertical-dist directory and modules chore: add horizontal-dist directory and modules
