+++ horizontal-dist/dockerfile
@@ -1,0 +1,5 @@
+FROM python:3.9-slim
+WORKDIR /app
+COPY horizontal_node.py /app/horizontal_node.py
+
+CMD ["python", "horizontal_node.py"]
