--- layer/dockerfile
@@ -1,5 +1,0 @@
-FROM python:3.9-slim
-WORKDIR /app
-COPY layer.py /app/layer.py
-
-CMD ["python", "layer.py"]
