+++ vertical-dist/dockerfile
@@ -1,0 +1,5 @@
+FROM python:3.9-slim
+WORKDIR /app
+COPY vertical_node.py /app/vertical_node.py
+
+CMD ["python", "vertical_node.py"]
