+++ vertical-dist/run_vertical_fcnn.py
