+++ vertical-dist/vertical_node.py
